pyproject.toml
[project]
name = "local_rag"
version = "0.1.0"
description = "Local RAG application for MCQ answering using ChromaDB and LangGraph"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "transformers>=4.40.0",
    "sentence-transformers>=2.7.0",
    "torch>=2.3.0",
    "chromadb>=0.5.0",
    "pymupdf>=1.24.0",
    "langgraph>=0.2.0",
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-huggingface>=0.0.3",
    "streamlit>=1.34.0",
    "python-dotenv>=1.0.0",
    "openai>=1.30.0",
    "anthropic>=0.26.0",
    "google-generativeai>=0.5.0",
    "huggingface-hub>=0.23.0",
    "accelerate>=1.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.poe.tasks.checks]
help = "Run linting, formatting and type checks on src/ folder."
shell = "uv run black src/ && uv run ruff check --fix src/ && uv run pyright src/"

[tool.poe.tasks.download]
help = "Download the HuggingFace embedding model (required for vector store)"
cmd = "uv run python -m src.pipelines.download_models"

[tool.poe.tasks.ingest]
help   = "Run the ingestion pipeline to ingest documents into the ChromaDB vectorstore"
cmd = "uv run python -m src.pipelines.ingest_documents"

[tool.poe.tasks.test]
help = "Run tests on sample questions using the RAG workflow; Note by default it uses OpenAI LLM and top 3 chunks"
cmd = "uv run python -m src.pipelines.test_questions openai 3"

[tool.poe.tasks.chat]
help = "Run the Streamlit Chatbot UI application"
cmd = "uv run streamlit run src/frontend/app.py"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "B008"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.pyright]
include = ["src"]
exclude = ["**/__pycache__"]
pythonVersion = "3.11"
typeCheckingMode = "basic"
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportUnknownVariableType = false

[dependency-groups]
dev = [
    "poethepoet>=0.37.0",
    "ruff>=0.4.0",
    "black>=24.4.0",
    "pyright>=1.1.360",
]
